+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/JavascriptFirebaseFunctionsExternals.kt
@@ -1,5 +1,6 @@
-package com.gchristov.thecodinglove.commonservicedata.api
+package com.gchristov.thecodinglove.commonservicedata
 
+import com.gchristov.thecodinglove.commonservicedata.api.FirebaseFunctionsHttps
 import com.gchristov.thecodinglove.commonservicedata.pubsub.FirebaseFunctionsPubSub
 
 external var exports: dynamic
