+++ appBackend/slack/src/jsMain/kotlin/com/gchristov/thecodinglove/slack/SlackSlashCommandApiService.kt
@@ -2,6 +2,7 @@ package com.gchristov.thecodinglove.slack
 
 import com.gchristov.thecodinglove.commonservice.ApiService
 import com.gchristov.thecodinglove.commonservicedata.api.*
+import com.gchristov.thecodinglove.commonservicedata.exports
 import com.gchristov.thecodinglove.slack.usecase.VerifySlackRequestUseCase
 import com.gchristov.thecodinglove.slackdata.api.ApiSlackSlashCommand
 import kotlinx.serialization.encodeToString
